+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionMap.cs
@@ -180,7 +180,7 @@ namespace UnityEngine.InputSystem
         /// </summary>
         /// <seealso cref="InputAction.started"/>
         /// <seealso cref="InputAction.performed"/>
-        /// <seealso cref="InputAction.cancelled"/>
+        /// <seealso cref="InputAction.canceled"/>
         public event Action<InputAction.CallbackContext> actionTriggered
         {
             add => m_ActionCallbacks.AppendWithCapacity(value);