+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Steam/SteamSupport.cs
@@ -8,7 +8,12 @@ namespace UnityEngine.InputSystem.Steam
     /// <summary>
     /// Adds support for Steam controllers.
     /// </summary>
-    public static class SteamSupport
+#if UNITY_DISABLE_DEFAULT_INPUT_PLUGIN_INITIALIZATION
+    public
+#else
+    internal
+#endif
+    static class SteamSupport
     {
         /// <summary>
         /// Wrapper around the Steam controller API.