+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XInput/XInputSupport.cs
@@ -7,7 +7,12 @@ namespace UnityEngine.InputSystem.XInput
     /// <summary>
     /// Adds support for XInput controllers.
     /// </summary>
-    public static class XInputSupport
+#if UNITY_DISABLE_DEFAULT_INPUT_PLUGIN_INITIALIZATION
+    public
+#else
+    internal
+#endif
+    static class XInputSupport
     {
         public static void Initialize()
         {