+++ Packages/com.unity.inputsystem/InputSystem/Plugins/iOS/iOSSupport.cs
@@ -3,7 +3,12 @@ using UnityEngine.InputSystem.Layouts;
 
 namespace UnityEngine.InputSystem.iOS
 {
-    public static class iOSSupport
+#if UNITY_DISABLE_DEFAULT_INPUT_PLUGIN_INITIALIZATION
+    public
+#else
+    internal
+#endif
+    static class iOSSupport
     {
         public static void Initialize()
         {